src/com/thoughtworks/domain/AccountTransaction.java
package com.thoughtworks.domain;

import java.math.BigDecimal;
import java.sql.Date;

public class AccountTransaction extends DomainObject {
    private Long accountID;
    private Date transactionDate;
    private String transactionType;
    private BigDecimal amount;

    public Long getAccountID() {
        return accountID;
    }

    public void setAccountID(Long accountID) {
        this.accountID = accountID;
    }

    public Date getTransactionDate() {
        return transactionDate;
    }

    public void setTransactionDate(Date transactionDate) {
        this.transactionDate = transactionDate;
    }

    public String getTransactionType() {
        return transactionType;
    }

    public void setTransactionType(String transactionType) {
        this.transactionType = transactionType;
    }

    public BigDecimal getAmount() {
        return amount;
    }

    public void setAmount(BigDecimal amount) {
        this.amount = amount;
    }
}
